Split AnyHarness session actor support modules
