+++ anyharness/crates/anyharness-lib/src/acp/session_actor/stderr.rs
@@ -1,0 +1,56 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(super) enum AgentStderrSeverity {
+    Error,
+    Warn,
+    Debug,
+}
+
+pub(super) fn sanitize_agent_stderr_line(line: &str) -> String {
+    strip_ansi_escape_codes(line).trim().to_string()
+}
+
+pub(super) fn classify_agent_stderr_line(line: &str) -> AgentStderrSeverity {
+    let upper = line.to_ascii_uppercase();
+
+    if has_log_level_token(&upper, "ERROR") {
+        AgentStderrSeverity::Error
+    } else if has_log_level_token(&upper, "WARN") || has_log_level_token(&upper, "WARNING") {
+        AgentStderrSeverity::Warn
+    } else if has_log_level_token(&upper, "INFO")
+        || has_log_level_token(&upper, "DEBUG")
+        || has_log_level_token(&upper, "TRACE")
+    {
+        AgentStderrSeverity::Debug
+    } else {
+        AgentStderrSeverity::Warn
+    }
+}
+
+fn has_log_level_token(line: &str, level: &str) -> bool {
+    line.starts_with(level)
+        || line.contains(&format!(" {level} "))
+        || line.contains(&format!(":{level} "))
+        || line.contains(&format!(" {level}:"))
+}
+
+fn strip_ansi_escape_codes(input: &str) -> String {
+    let mut cleaned = String::with_capacity(input.len());
+    let mut chars = input.chars().peekable();
+    while let Some(ch) = chars.next() {
+        if ch == '\u{1b}' {
+            if matches!(chars.peek(), Some('[')) {
+                chars.next();
+                for escape_char in chars.by_ref() {
+                    if ('@'..='~').contains(&escape_char) {
+                        break;
+                    }
+                }
+            }
+            continue;
+        }
+
+        cleaned.push(ch);
+    }
+
+    cleaned
+}
